parser-api/requirements.txt
Flask==3.0.0
gunicorn==21.2.0
google-cloud-storage==2.14.0
